+++ include/qpdf/Constants.h
@@ -83,14 +83,15 @@ enum qpdf_exit_code_e {
 
 enum qpdf_error_code_e {
     qpdf_e_success = 0,
-    qpdf_e_internal,    /* logic/programming error -- indicates bug */
-    qpdf_e_system,      /* I/O error, memory error, etc. */
-    qpdf_e_unsupported, /* PDF feature not (yet) supported by qpdf */
-    qpdf_e_password,    /* incorrect password for encrypted file */
-    qpdf_e_damaged_pdf, /* syntax errors or other damage in PDF */
-    qpdf_e_pages,       /* erroneous or unsupported pages structure */
-    qpdf_e_object,      /* type/bounds errors accessing objects */
-    qpdf_e_json,        /* error in qpdf JSON */
+    qpdf_e_internal,      /* logic/programming error -- indicates bug */
+    qpdf_e_system,        /* I/O error, memory error, etc. */
+    qpdf_e_unsupported,   /* PDF feature not (yet) supported by qpdf */
+    qpdf_e_password,      /* incorrect password for encrypted file */
+    qpdf_e_damaged_pdf,   /* syntax errors or other damage in PDF */
+    qpdf_e_pages,         /* erroneous or unsupported pages structure */
+    qpdf_e_object,        /* type/bounds errors accessing objects */
+    qpdf_e_json,          /* error in qpdf JSON */
+    qpdf_e_linearization, /* linearization warning */
 };
 
 /* Object Types */